Разгрузил инит, разделил обработчики
