+++ cartoonfilter_ds_project/handlers.py
@@ -4,9 +4,9 @@ from telegram.ext import Updater, CommandHandler, MessageHandler
 from telegram import ReplyKeyboardMarkup, utils
 import urllib.request as ur
 
-from filters import cartoonize_using_network_Blue as nb
+from neuro_filters import cartoonize_using_network_Blue as nb
 from filters import cartoonise_using_cartoonfilter as cf
-from filters import cartoonize_using_network_without_filters as wf
+from neuro_filters import cartoonize_using_network_without_filters as wf
 
 def main_keyboard():
     return ReplyKeyboardMarkup([
