+++ cartoonfilter_ds_project/webapp/config.py
@@ -3,7 +3,7 @@ import os
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
-SECRET_KEY = 'Your secret key'
+SECRET_KEY = ''
 ALLOWED_EXTENSIONS = {'jpg'}
 SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, '..', 'webapp.db')
